requirements.txt
fastapi
uvicorn
python-dotenv
requests
pymssql